Refactor analyses toolbar.
